+++ packages/calcite-components/src/components/input-date-picker/input-date-picker.tsx
@@ -131,7 +131,7 @@ export class InputDatePicker
   }
 
   /**
-   * The ID of the form that will be associated with the component.
+   * The `id` of the form that will be associated with the component.
    *
    * When not set, the component will be associated with its ancestor form element, if any.
    */
@@ -274,7 +274,7 @@ export class InputDatePicker
   @Prop() validationMessage: string;
 
   /** Specifies the validation icon to display under the component. */
-  @Prop() validationIcon: string | boolean;
+  @Prop({ reflect: true }) validationIcon: string | boolean;
 
   /**
    * Specifies the name of the component.
@@ -379,7 +379,7 @@ export class InputDatePicker
   //--------------------------------------------------------------------------
 
   /**
-   * Fires when the component's value changes.
+   * Fires when the component's `value` changes.
    */
   @Event({ cancelable: false }) calciteInputDatePickerChange: EventEmitter<void>;