+++ packages/calcite-components-angular/projects/component-library/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.0-next.17](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.2.0-next.16...@esri/calcite-components-angular@2.2.0-next.17) (2024-01-11)
+
+**Note:** Version bump only for package @esri/calcite-components-angular
+
+## [2.2.0-next.16](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.2.0-next.15...@esri/calcite-components-angular@2.2.0-next.16) (2024-01-10)
+
+**Note:** Version bump only for package @esri/calcite-components-angular
+
 ## [2.2.0-next.15](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.2.0-next.14...@esri/calcite-components-angular@2.2.0-next.15) (2024-01-10)
 
 **Note:** Version bump only for package @esri/calcite-components-angular